requirements.txt
flake8==3.0.4
flake8-double-quotes==0.0.1
flake8-import-order==0.9.2
mccabe==0.5.2
pep8-naming==0.4.1
pycodestyle==2.0.0
pyflakes==1.2.3
six==1.10.0
django==1.10.2
django-filter==0.15.2
djangorestframework==3.4.7
celery==4.0.0
PyYAML==3.12
